drivers: usb_dc_rpi_pico: correctly toggle between data1/data0 on to_device control transfers
The data stage of Control transfers that are sent from Host to Device, can be made out of
multiple OUT transactions, if the amount of data to be transmitted is larger than the endpoint
size. When this happens, the DATA pid should be toggling. The USB Device driver of the pico
must correctly prime the EP0_OUT buffer with the correct data PID, otherwise the hardware will
reject the received transaction.

Before this change the driver used to always prime EP0_OUT with a DATA1 pid. After this change
the driver only uses DATA1 pid after a setup transaction, and then toggles the pid for each transaction.

Signed-off-by: Purdea Andrei <andrei@purdea.ro>